+++ oo_patterns/__init__.py
@@ -15,7 +15,7 @@ DESCRIPTION = (
 
 VERSION = (
     0,
-    4,
+    5,
     0,
 )
 PY_VERSION = (