+++ oo_patterns/tests/behavioral/observer/test_publishers.py
@@ -1,20 +1,32 @@
+__all__ = (
+    "PublisherTestCase",
+    "AsyncPublisherTestCase",
+)
+
 import typing as t
 from unittest import IsolatedAsyncioTestCase, TestCase
 from unittest.mock import AsyncMock, Mock
 
 from oo_patterns.behavioral.observer.publishers import AsyncPublisher, Publisher
+from oo_patterns.tests.helpers import TestingMixin
 
 if t.TYPE_CHECKING:
     ...
 
 
-class PublisherInterfaceTestCase(TestCase):
+PublisherTypeVar = t.TypeVar("PublisherTypeVar", bound=Publisher)
+
+
+class PublisherTestCase(
+    TestingMixin[PublisherTypeVar], TestCase, t.Generic[PublisherTypeVar]
+):
+    tst_cls = Publisher
 
     def test_main_flow(self):
         mock_subscriber_one = Mock()
         mock_subscriber_two = Mock(side_effect=Exception("Test"))
 
-        tst_obj: "Publisher" = Publisher()
+        tst_obj = self.build_tst_obj()
         tst_obj.add_subscribers(mock_subscriber_one, mock_subscriber_two)
         tst_obj.add_subscribers(mock_subscriber_one)
 
@@ -42,14 +54,24 @@ class PublisherInterfaceTestCase(TestCase):
         mock_subscriber_one.assert_not_called()
         mock_subscriber_two.assert_not_called()
 
+        tst_obj.remove_all_subscribers()
 
-class AsyncPublisherInterfaceTestCase(IsolatedAsyncioTestCase):
+
+AsyncPublisherTypeVar = t.TypeVar("AsyncPublisherTypeVar", bound=AsyncPublisher)
+
+
+class AsyncPublisherTestCase(
+    TestingMixin[AsyncPublisherTypeVar],
+    IsolatedAsyncioTestCase,
+    t.Generic[AsyncPublisherTypeVar],
+):
+    tst_cls = AsyncPublisher
 
     async def test_main_flow(self):
         mock_subscriber_one = AsyncMock()
         mock_subscriber_two = AsyncMock(side_effect=Exception("Test"))
 
-        tst_obj: "AsyncPublisher" = AsyncPublisher()
+        tst_obj = self.build_tst_obj()
         tst_obj.add_subscribers(mock_subscriber_one, mock_subscriber_two)
         tst_obj.add_subscribers(mock_subscriber_one)
 
@@ -76,3 +98,5 @@ class AsyncPublisherInterfaceTestCase(IsolatedAsyncioTestCase):
         await tst_obj.notify_subscribers(context=Mock())
         mock_subscriber_one.assert_not_awaited()
         mock_subscriber_two.assert_not_awaited()
+
+        tst_obj.remove_all_subscribers()