+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# OOP Changelog
+# OO Patterns Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2024-07-27
+
+### Added
+- Implement `TestingMixin` helper for unifying testing processes. See `oo_patterns/tests/helpers.py` file for details.
+- Add `remove_all_subscribers` method into `Observer` classes. See `oo_patterns/behavioral/observer/` directory for details.
 
 ## [0.4.0] - 2024-07-21